backend/requirements.txt
flask
flask-bcrypt
flask-login
flask-jwt-extended
flask-cors
sqlalchemy
pymongo
pillow